+++ packages/BoPlugin/index.js
@@ -1,0 +1,9 @@
+import {
+    watermark
+} from './src/watermark.js';
+
+export default {
+    install(Vue) {
+        Vue.prototype.$watermark = watermark
+    }
+}